Create Task-1B
T A S K 1 B - Naïve Bayesian Classifier for Text Classification

(Assuming a set of documents that need to be classified, use the
naïve Bayesian classifier model to perform this task. Built-in Java classes/API
can be used to write the program. Calculate the accuracy, precision, and recall
for your dataset.)